refactor(lib): Cleanup linter errors
