+++ example.js
@@ -2,7 +2,7 @@
 const Observable = require('rxjs/Observable').Observable;
 const logSymbols = require('log-symbols');
 const delay = require('delay');
-const Listr = require('./');
+const Listr = require('.');
 
 const renderer = process.argv[2];
