+++ index.js
@@ -14,7 +14,6 @@ const runTask = (task, context, errors) => {
 };
 
 class Listr {
-
 	constructor(tasks, opts) {
 		if (tasks && !Array.isArray(tasks) && typeof tasks === 'object') {
 			if (typeof tasks.title === 'string' && typeof tasks.task === 'function') {
