+++ lib/task-wrapper.js
@@ -3,7 +3,6 @@ const state = require('./state');
 const ListrError = require('./listr-error');
 
 class TaskWrapper {
-
 	constructor(task, errors) {
 		this._task = task;
 		this._errors = errors;
