+++ test/skip.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import Listr from '../';
+import Listr from '..';
 import SimpleRenderer from './fixtures/simple-renderer';
 import {testOutput} from './fixtures/utils';
