+++ test/stream.js
@@ -2,7 +2,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import test from 'ava';
 import split from 'split';
-import Listr from '../';
+import Listr from '..';
 import SimpleRenderer from './fixtures/simple-renderer';
 import {testOutput} from './fixtures/utils';
